Extract macros into its own crate
