+++ Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 members = [
     "components/glsl-watcher",
+    "components/macros",
 ]
 
 [package]
@@ -21,6 +22,7 @@ glsl-watcher = { path = "components/glsl-watcher", version = "0.0.1" }
 glutin = "0.29"
 log = "0.4"
 notify = "4.0.12"
+macros = { path = "components/macros", version = "0.1" }
 memoffset = "0.6"
 regex = "1.5.4"
 serde = { version = "1.0", features = ["derive"] }
