+++ components/macros/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "macros"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+glow = "0.11"
+tracing = { version = "0.1", default-features = false, features = ["std"] }
