+++ src/lib.rs
@@ -6,7 +6,6 @@ pub mod app;
 pub mod config;
 pub mod event;
 pub mod input;
-pub mod macros;
 pub mod minime;
 pub mod mouse;
 pub mod render;
@@ -22,7 +21,6 @@ pub use app::*;
 pub use config::*;
 pub use event::*;
 pub use input::*;
-pub use macros::*;
 pub use minime::*;
 pub use mouse::*;
 pub use render::*;
