+++ src/ui/mod.rs
@@ -23,43 +23,3 @@ mod misc_util;
 mod post_process;
 mod shader_version;
 pub mod vao;
-
-/// Check for OpenGL error and report it using `tracing::error`.
-///
-/// WARNING: slow! Only use during setup!
-///
-/// ``` no_run
-/// # let glow_context = todo!();
-/// use egui_glow::check_for_gl_error_even_in_release;
-/// check_for_gl_error_even_in_release!(glow_context);
-/// check_for_gl_error_even_in_release!(glow_context, "during painting");
-/// ```
-#[macro_export]
-macro_rules! check_for_gl_error_even_in_release {
-    ($gl: expr) => {{
-        $crate::check_for_gl_error_impl($gl, file!(), line!(), "")
-    }};
-    ($gl: expr, $context: literal) => {{
-        $crate::check_for_gl_error_impl($gl, file!(), line!(), $context)
-    }};
-}
-
-/// Profiling macro for feature "puffin"
-macro_rules! profile_function {
-    ($($arg: tt)*) => {
-        #[cfg(feature = "puffin")]
-        #[cfg(not(target_arch = "wasm32"))]
-        puffin::profile_function!($($arg)*);
-    };
-}
-pub(crate) use profile_function;
-
-/// Profiling macro for feature "puffin"
-macro_rules! profile_scope {
-    ($($arg: tt)*) => {
-        #[cfg(feature = "puffin")]
-        #[cfg(not(target_arch = "wasm32"))]
-        puffin::profile_scope!($($arg)*);
-    };
-}
-pub(crate) use profile_scope;
